Fixing potential bad order of template loading and map construction
